validation: tm: use tm start and stop API
Add tm start API after creating a TM system and TM stop API before
destroying it. Also stop TM system before shaper/scheduler/threshold
or wred profile updates if dynamic update is not supported by the TM
system.

+++ test/validation/api/traffic_mngr/traffic_mngr.c
@@ -1664,6 +1664,14 @@ static int create_tm_system(void)
 		return -1;
 	}
 
+	/* Start TM system */
+	CU_ASSERT((rc = odp_tm_start(odp_tm)) == 0);
+	if (rc != 0) {
+		ODPH_ERR("odp_tm_start() failed for tm: %" PRIx64 "\n",
+			 odp_tm_to_u64(odp_tm));
+		return -1;
+	}
+
 	return 0;
 }
 
@@ -2042,11 +2050,16 @@ static int destroy_tm_systems(void)
 
 	/* Close/free the TM systems. */
 	for (idx = 0; idx < num_odp_tm_systems; idx++) {
+		if (odp_tm_stop(odp_tm_systems[idx]) != 0)
+			return -1;
+
 		if (destroy_tm_subtree(root_node_descs[idx]) != 0)
 			return -1;
 
 		if (odp_tm_destroy(odp_tm_systems[idx]) != 0)
 			return -1;
+
+		odp_tm_systems[idx] = ODP_TM_INVALID;
 	}
 
 	/* Close/free the TM profiles. */
@@ -2415,6 +2428,7 @@ static int set_shaper(const char    *node_name,
 	odp_tm_shaper_params_t shaper_params;
 	odp_tm_shaper_t        shaper_profile;
 	odp_tm_node_t          tm_node;
+	int rc;
 
 	tm_node = find_tm_node(0, node_name);
 	if (tm_node == ODP_TM_INVALID) {
@@ -2431,6 +2445,13 @@ static int set_shaper(const char    *node_name,
 	shaper_params.shaper_len_adjust = 0;
 	shaper_params.dual_rate         = 0;
 
+	if (!tm_capabilities.dynamic_shaper_update) {
+		/* Stop TM system before update when dynamic update is not
+		 * supported.
+		 */
+		CU_ASSERT_FATAL(odp_tm_stop(odp_tm_systems[0]) == 0);
+	}
+
 	/* First see if a shaper profile already exists with this name, in
 	 * which case we use that profile, else create a new one. */
 	shaper_profile = odp_tm_shaper_lookup(shaper_name);
@@ -2443,7 +2464,13 @@ static int set_shaper(const char    *node_name,
 		num_shaper_profiles++;
 	}
 
-	return odp_tm_node_shaper_config(tm_node, shaper_profile);
+	rc = odp_tm_node_shaper_config(tm_node, shaper_profile);
+
+	if (!tm_capabilities.dynamic_shaper_update) {
+		/* Start TM system, post update */
+		CU_ASSERT_FATAL(odp_tm_start(odp_tm_systems[0]) == 0);
+	}
+	return rc;
 }
 
 static int traffic_mngr_check_shaper(void)
@@ -2617,6 +2644,13 @@ static int set_sched_fanin(const char         *node_name,
 	if (node_desc == NULL)
 		return -1;
 
+	if (!tm_capabilities.dynamic_sched_update) {
+		/* Stop TM system before update when dynamic update is not
+		 * supported.
+		 */
+		CU_ASSERT_FATAL(odp_tm_stop(odp_tm_systems[0]) == 0);
+	}
+
 	fanin_cnt = MIN(node_desc->num_children, FANIN_RATIO);
 	for (fanin = 0; fanin < fanin_cnt; fanin++) {
 		odp_tm_sched_params_init(&sched_params);
@@ -2653,10 +2687,15 @@ static int set_sched_fanin(const char         *node_name,
 		rc = odp_tm_node_sched_config(tm_node, fanin_node,
 					      sched_profile);
 		if (rc != 0)
-			return -1;
+			goto exit;
 	}
 
-	return 0;
+exit:
+	if (!tm_capabilities.dynamic_sched_update) {
+		/* Start TM system, post update */
+		CU_ASSERT_FATAL(odp_tm_start(odp_tm_systems[0]) == 0);
+	}
+	return rc;
 }
 
 static int test_sched_queue_priority(const char *shaper_name,
@@ -2935,6 +2974,13 @@ static int set_queue_thresholds(odp_tm_queue_t             tm_queue,
 	odp_tm_threshold_t threshold_profile;
 	int ret;
 
+	if (!tm_capabilities.dynamic_threshold_update) {
+		/* Stop TM system before update when dynamic update is not
+		 * supported.
+		 */
+		CU_ASSERT_FATAL(odp_tm_stop(odp_tm_systems[0]) == 0);
+	}
+
 	/* First see if a threshold profile already exists with this name, in
 	 * which case we use that profile, else create a new one. */
 	threshold_profile = odp_tm_thresholds_lookup(threshold_name);
@@ -2942,17 +2988,25 @@ static int set_queue_thresholds(odp_tm_queue_t             tm_queue,
 		ret = odp_tm_thresholds_params_update(threshold_profile,
 						      threshold_params);
 		if (ret)
-			return ret;
+			goto exit;
 	} else {
 		threshold_profile = odp_tm_threshold_create(threshold_name,
 							    threshold_params);
-		if (threshold_profile == ODP_TM_INVALID)
-			return -1;
+		if (threshold_profile == ODP_TM_INVALID) {
+			ret = -1;
+			goto exit;
+		}
 		threshold_profiles[num_threshold_profiles] = threshold_profile;
 		num_threshold_profiles++;
 	}
 
-	return odp_tm_queue_threshold_config(tm_queue, threshold_profile);
+	ret = odp_tm_queue_threshold_config(tm_queue, threshold_profile);
+exit:
+	if (!tm_capabilities.dynamic_threshold_update) {
+		/* Start TM system, post update */
+		CU_ASSERT_FATAL(odp_tm_start(odp_tm_systems[0]) == 0);
+	}
+	return ret;
 }
 
 static int test_threshold(const char *threshold_name,
@@ -3049,6 +3103,7 @@ static int set_queue_wred(odp_tm_queue_t   tm_queue,
 {
 	odp_tm_wred_params_t wred_params;
 	odp_tm_wred_t        wred_profile;
+	int rc;
 
 	odp_tm_wred_params_init(&wred_params);
 	if (use_dual_slope) {
@@ -3065,6 +3120,13 @@ static int set_queue_wred(odp_tm_queue_t   tm_queue,
 
 	wred_params.enable_wred       = true;
 	wred_params.use_byte_fullness = use_byte_fullness;
+
+	if (!tm_capabilities.dynamic_wred_update) {
+		/* Stop TM system before update when dynamic update is not
+		 * supported.
+		 */
+		CU_ASSERT_FATAL(odp_tm_stop(odp_tm_systems[0]) == 0);
+	}
 
 	/* First see if a wred profile already exists with this name, in
 	 * which case we use that profile, else create a new one. */
@@ -3084,7 +3146,14 @@ static int set_queue_wred(odp_tm_queue_t   tm_queue,
 		}
 	}
 
-	return odp_tm_queue_wred_config(tm_queue, pkt_color, wred_profile);
+	rc = odp_tm_queue_wred_config(tm_queue, pkt_color, wred_profile);
+
+	if (!tm_capabilities.dynamic_wred_update) {
+		/* Start TM system, post update */
+		CU_ASSERT_FATAL(odp_tm_start(odp_tm_systems[0]) == 0);
+	}
+	return rc;
+
 }
 
 static int test_byte_wred(const char      *wred_name,
